Refactor: utils.js에 있던 navigateTo 함수를 router.js로 이전
