+++ src/router.js
@@ -1,3 +1,4 @@
+import { autoSaveDocument, manualSaveDocument } from "./editor.js";
 import { renderEditor, renderSidebar } from "./rendering.js";
 import { fetchDocumentContent, fetchDocuments } from "./utils.js";
 
@@ -7,7 +8,9 @@ const render = async (docId = "") => {
   if (pathname === "/") {
     document.getElementById("doc-title__input").value = `🥔 감자의 Notion`;
     document.getElementById("doc__title").innerText = `🥔 감자의 Notion`;
-    document.getElementById("doc-contents").value = `🥔 감자의 Notion에 오신 것을 환영합니다!
+    document.getElementById(
+      "doc-contents"
+    ).value = `🥔 감자의 Notion에 오신 것을 환영합니다!
 작성한 문서를 확인해보세요! 새로운 문서를 추가하거나 기존 문서를 삭제하는 것도 가능합니다.
     `;
     const documents = await fetchDocuments();
@@ -20,8 +23,34 @@ const render = async (docId = "") => {
   }
 };
 
+// 페이지를 렌더링하는 함수
+export const navigateTo = async (state = { id: null }, pathname) => {
+  history.pushState(state, null, pathname);
+
+  if (pathname === "/") {
+    render(state.id);
+  } else {
+    const documentContent = await fetchDocumentContent(state.id);
+    renderEditor(documentContent);
+
+    autoSaveDocument(state.id);
+    manualSaveDocument(state.id);
+  }
+};
+
 // 페이지 로드 시 라우터 실행
 document.addEventListener("DOMContentLoaded", render);
+
+document.body.addEventListener("click", (e) => {
+  e.preventDefault();
+  const target = e.target;
+  const id = target.dataset.id;
+
+  if (target.tagName === "A") {
+    const pathname = new URL(target.href).pathname;
+    navigateTo({ id }, pathname);
+  }
+});
 
 // popstate 이벤트에서 현재 경로를 전달하여 렌더링
 window.addEventListener("popstate", (e) => {
