+++ src/utils.js
@@ -1,6 +1,3 @@
-import { autoSaveDocument, manualSaveDocument } from "./editor.js";
-import { renderEditor } from "./rendering.js";
-
 const BASE_URL = `https://kdt-api.fe.dev-cos.com/documents`;
 const username = `potatoes`;
 
@@ -39,17 +36,6 @@ export const fetchDocumentContent = async (docId = "") => {
   } catch (error) {
     console.error("문서 내용 요청 실패:", error);
   }
-};
-
-// 페이지를 렌더링하는 함수
-export const navigateTo = async (state = { id: null }, pathname) => {
-  history.pushState(state, null, pathname);
-
-  const documentContent = await fetchDocumentContent(state.id);
-  renderEditor(documentContent);
-
-  autoSaveDocument(state.id);
-  manualSaveDocument(state.id);
 };
 
 export const fetchDeleteDocument = async (docId = "") => {
